+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - `Innmind\IO\Sockets\Servers`
 - `Innmind\IO\Sockets\Server\Pool`
 
+### Changed
+
+- Requires `innmind/immutable:~5.2`
+
 ## 2.6.0 - 2024-03-09
 
 ### Added